TST: consolidated the UniFrac tests, since we're dropping the edu versions
